task completed

+++ src/main/java/core/basesyntax/strategy/impl/DefaultDiscountService.java
@@ -3,9 +3,8 @@ package core.basesyntax.strategy.impl;
 import core.basesyntax.strategy.DiscountService;
 
 public class DefaultDiscountService implements DiscountService {
-
     @Override
     public double getDiscount() {
-        return (double) 0.0;
+        return 0.0;
     }
 }
